+++ inlang/source-code/cli/src/utilities/getInlangProject.ts
@@ -3,39 +3,36 @@ import { loadProject, type InlangProject } from "@inlang/sdk"
 import { resolve } from "node:path"
 import { openRepository, findRepoRoot } from "@lix-js/client"
 import { id } from "../../marketplace-manifest.json"
-import { telemetry } from "../services/telemetry/index.js"
 
 /**
- * Gets the inlang project and exists if the project contains errors.
+ * Used for telemetry.
+ */
+export let lastUsedProject: InlangProject | undefined
+
+/**
+ * Gets the inlang project and exits if the project contains errors.
  */
 export async function getInlangProject(args: { projectPath: string }): Promise<InlangProject> {
 	const baseDirectory = process.cwd()
 	const projectPath = resolve(baseDirectory, args.projectPath)
 
-	const repoRoot = await findRepoRoot({ nodeishFs: fs, path: projectPath })
+	let repoRoot = await findRepoRoot({ nodeishFs: fs, path: projectPath })
 
-	let project
 	if (!repoRoot) {
 		console.error(
 			`Could not find repository root for path ${projectPath}, falling back to direct fs access`
 		)
-
-		project = await loadProject({
-			projectPath,
-			nodeishFs: fs,
-			appId: id,
-		})
-	} else {
-		const repo = await openRepository(repoRoot, {
-			nodeishFs: fs,
-		})
-
-		project = await loadProject({
-			projectPath,
-			repo,
-			appId: id,
-		})
+		repoRoot = baseDirectory
 	}
+	const repo = await openRepository(repoRoot, {
+		nodeishFs: fs,
+	})
+
+	const project = await loadProject({
+		projectPath,
+		repo,
+		appId: id,
+	})
 
 	if (project.errors().length > 0) {
 		for (const error of project.errors()) {
@@ -43,11 +40,6 @@ export async function getInlangProject(args: { projectPath: string }): Promise<I
 		}
 		process.exit(1)
 	}
-	if (project.id) {
-		telemetry.groupIdentify({
-			groupType: "project",
-			groupKey: project.id,
-		})
-	}
+	lastUsedProject = project
 	return project
 }